fix PR comment
